final commit

+++ frontend/src/components/WorkoutDetails.js
@@ -1,15 +1,27 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutContext } from "../hooks/useWorkoutContext"
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
+
+
+
 const WorkoutDetails = ({workout}) => {
 
     const {dispatch} = useWorkoutContext();
+    const {user} = useAuthContext();
 
     const handleDelete = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`,
+            }
         })
 
         const json = await response.json()
